refactor(columnar): extract merge operators
