+++ tests/meson.build
@@ -819,6 +819,7 @@ backend_src = files(
   '../wirelog/columnar/expr.c',
   '../wirelog/columnar/filter.c',
   '../wirelog/columnar/join.c',
+  '../wirelog/columnar/merge.c',
   '../wirelog/columnar/eval_stack.c',
   '../wirelog/columnar/eval.c',
   '../wirelog/columnar/arrangement.c',
